Move more fields from the common TopicConfig into TopicConsumerConfig along with the corresponding subclasses.
