+++ data-prepper-plugins/kafka-plugins/src/main/java/org/opensearch/dataprepper/plugins/kafka/buffer/BufferTopicConfig.java
@@ -28,6 +28,13 @@ class BufferTopicConfig extends CommonTopicConfig implements TopicProducerConfig
     static final Duration DEFAULT_FETCH_MAX_WAIT = Duration.ofMillis(500);
     static final ByteCount DEFAULT_FETCH_MIN_BYTES = ByteCount.parse("1b");
     static final ByteCount DEFAULT_MAX_PARTITION_FETCH_BYTES = ByteCount.parse("1mb");
+    static final Duration DEFAULT_SESSION_TIMEOUT = Duration.ofSeconds(45);
+    static final String DEFAULT_AUTO_OFFSET_RESET = "earliest";
+    static final Duration DEFAULT_THREAD_WAITING_TIME = Duration.ofSeconds(5);
+    static final Duration DEFAULT_MAX_POLL_INTERVAL = Duration.ofSeconds(300);
+    static final Integer DEFAULT_CONSUMER_MAX_POLL_RECORDS = 500;
+    static final Integer DEFAULT_NUM_OF_WORKERS = 2;
+    static final Duration DEFAULT_HEART_BEAT_INTERVAL_DURATION = Duration.ofSeconds(5);
 
 
     @JsonProperty("encryption_key")
@@ -52,6 +59,38 @@ class BufferTopicConfig extends CommonTopicConfig implements TopicProducerConfig
 
     @JsonProperty("create_topic")
     private boolean isCreateTopic = false;
+
+    @JsonProperty("group_id")
+    @Valid
+    @Size(min = 1, max = 255, message = "size of group id should be between 1 and 255")
+    private String groupId;
+
+    @JsonProperty("workers")
+    @Valid
+    @Size(min = 1, max = 200, message = "Number of worker threads should lies between 1 and 200")
+    private Integer workers = DEFAULT_NUM_OF_WORKERS;
+
+    @JsonProperty("session_timeout")
+    @Valid
+    @Size(min = 1)
+    private Duration sessionTimeOut = DEFAULT_SESSION_TIMEOUT;
+
+    @JsonProperty("auto_offset_reset")
+    private String autoOffsetReset = DEFAULT_AUTO_OFFSET_RESET;
+
+    @JsonProperty("thread_waiting_time")
+    private Duration threadWaitingTime = DEFAULT_THREAD_WAITING_TIME;
+
+    @JsonProperty("max_poll_interval")
+    private Duration maxPollInterval = DEFAULT_MAX_POLL_INTERVAL;
+
+    @JsonProperty("consumer_max_poll_records")
+    private Integer consumerMaxPollRecords = DEFAULT_CONSUMER_MAX_POLL_RECORDS;
+
+    @JsonProperty("heart_beat_interval")
+    @Valid
+    @Size(min = 1)
+    private Duration heartBeatInterval= DEFAULT_HEART_BEAT_INTERVAL_DURATION;
 
     @JsonProperty("auto_commit")
     private Boolean autoCommit = DEFAULT_AUTO_COMMIT;
@@ -90,6 +129,11 @@ class BufferTopicConfig extends CommonTopicConfig implements TopicProducerConfig
     }
 
     @Override
+    public String getGroupId() {
+        return groupId;
+    }
+
+    @Override
     public Duration getCommitInterval() {
         return commitInterval;
     }
@@ -118,6 +162,41 @@ class BufferTopicConfig extends CommonTopicConfig implements TopicProducerConfig
     @Override
     public Boolean getAutoCommit() {
         return autoCommit;
+    }
+
+    @Override
+    public Duration getSessionTimeOut() {
+        return sessionTimeOut;
+    }
+
+    @Override
+    public String getAutoOffsetReset() {
+        return autoOffsetReset;
+    }
+
+    @Override
+    public Duration getThreadWaitingTime() {
+        return threadWaitingTime;
+    }
+
+    @Override
+    public Duration getMaxPollInterval() {
+        return maxPollInterval;
+    }
+
+    @Override
+    public Integer getConsumerMaxPollRecords() {
+        return consumerMaxPollRecords;
+    }
+
+    @Override
+    public Integer getWorkers() {
+        return workers;
+    }
+
+    @Override
+    public Duration getHeartBeatInterval() {
+        return heartBeatInterval;
     }
 
     @Override
