+++ data-prepper-plugins/kafka-plugins/src/main/java/org/opensearch/dataprepper/plugins/kafka/source/SourceTopicConfig.java
@@ -24,6 +24,13 @@ class SourceTopicConfig extends CommonTopicConfig implements TopicConsumerConfig
     static final Integer DEFAULT_FETCH_MAX_WAIT = 500;
     static final String DEFAULT_FETCH_MIN_BYTES = "1b";
     static final String DEFAULT_MAX_PARTITION_FETCH_BYTES = "1mb";
+    static final Duration DEFAULT_SESSION_TIMEOUT = Duration.ofSeconds(45);
+    static final String DEFAULT_AUTO_OFFSET_RESET = "earliest";
+    static final Duration DEFAULT_THREAD_WAITING_TIME = Duration.ofSeconds(5);
+    static final Duration DEFAULT_MAX_POLL_INTERVAL = Duration.ofSeconds(300);
+    static final Integer DEFAULT_CONSUMER_MAX_POLL_RECORDS = 500;
+    static final Integer DEFAULT_NUM_OF_WORKERS = 2;
+    static final Duration DEFAULT_HEART_BEAT_INTERVAL_DURATION = Duration.ofSeconds(5);
 
 
     @JsonProperty("serde_format")
@@ -36,6 +43,38 @@ class SourceTopicConfig extends CommonTopicConfig implements TopicConsumerConfig
 
     @JsonProperty("key_mode")
     private KafkaKeyMode kafkaKeyMode = KafkaKeyMode.INCLUDE_AS_FIELD;
+
+    @JsonProperty("group_id")
+    @Valid
+    @Size(min = 1, max = 255, message = "size of group id should be between 1 and 255")
+    private String groupId;
+
+    @JsonProperty("workers")
+    @Valid
+    @Size(min = 1, max = 200, message = "Number of worker threads should lies between 1 and 200")
+    private Integer workers = DEFAULT_NUM_OF_WORKERS;
+
+    @JsonProperty("session_timeout")
+    @Valid
+    @Size(min = 1)
+    private Duration sessionTimeOut = DEFAULT_SESSION_TIMEOUT;
+
+    @JsonProperty("auto_offset_reset")
+    private String autoOffsetReset = DEFAULT_AUTO_OFFSET_RESET;
+
+    @JsonProperty("thread_waiting_time")
+    private Duration threadWaitingTime = DEFAULT_THREAD_WAITING_TIME;
+
+    @JsonProperty("max_poll_interval")
+    private Duration maxPollInterval = DEFAULT_MAX_POLL_INTERVAL;
+
+    @JsonProperty("consumer_max_poll_records")
+    private Integer consumerMaxPollRecords = DEFAULT_CONSUMER_MAX_POLL_RECORDS;
+
+    @JsonProperty("heart_beat_interval")
+    @Valid
+    @Size(min = 1)
+    private Duration heartBeatInterval= DEFAULT_HEART_BEAT_INTERVAL_DURATION;
 
     @JsonProperty("auto_commit")
     private Boolean autoCommit = DEFAULT_AUTO_COMMIT;
@@ -78,6 +117,11 @@ class SourceTopicConfig extends CommonTopicConfig implements TopicConsumerConfig
     }
 
     @Override
+    public String getGroupId() {
+        return groupId;
+    }
+
+    @Override
     public MessageFormat getSerdeFormat() {
         return serdeFormat;
     }
@@ -117,5 +161,40 @@ class SourceTopicConfig extends CommonTopicConfig implements TopicConsumerConfig
     @Override
     public long getMaxPartitionFetchBytes() {
         return ByteCount.parse(maxPartitionFetchBytes).getBytes();
+    }
+
+    @Override
+    public Duration getSessionTimeOut() {
+        return sessionTimeOut;
+    }
+
+    @Override
+    public String getAutoOffsetReset() {
+        return autoOffsetReset;
+    }
+
+    @Override
+    public Duration getThreadWaitingTime() {
+        return threadWaitingTime;
+    }
+
+    @Override
+    public Duration getMaxPollInterval() {
+        return maxPollInterval;
+    }
+
+    @Override
+    public Integer getConsumerMaxPollRecords() {
+        return consumerMaxPollRecords;
+    }
+
+    @Override
+    public Integer getWorkers() {
+        return workers;
+    }
+
+    @Override
+    public Duration getHeartBeatInterval() {
+        return heartBeatInterval;
     }
 }
